renamed claimer to grabber
